Ed3.3

+++ CAPSProj/src/main/java/com/team6/CAPSProj/controller/AdminController.java
@@ -50,15 +50,20 @@ public class AdminController {
 
 	@RequestMapping(value = "/studentlist")
 
-	public String listStudent(Model model, HttpSession session) {
+	public String listStudent(Model model, HttpSession session, Boolean delete_student_status) {
 		Admin ad = (Admin) session.getAttribute("usession");
 		if (ad == null)
 		{
 			return "redirect:/home";
 		}
-		else {
-		model.addAttribute("studentlist", stservice.findAllStudents());
-		return "admin_student_manage";
+		else 
+		{
+			model.addAttribute("studentlist", stservice.findAllStudents());
+			
+			if(delete_student_status != null)
+				model.addAttribute("status", delete_student_status);
+			
+			return "admin_student_manage";
 		}
 	}
 	
@@ -119,7 +124,7 @@ public class AdminController {
 	}
 	
 	@RequestMapping(value = "/deletestudent/{matricNo}")
-	public String deleteStudent(@PathVariable("matricNo") String matricNo, HttpSession session) {
+	public String deleteStudent(@PathVariable("matricNo") String matricNo, HttpSession session, Model model) {
 		Admin ad = (Admin) session.getAttribute("usession");
 		if (ad == null)
 		{
@@ -131,34 +136,51 @@ public class AdminController {
 		Student student = stservice.findStudentByMatricNo(matricNo);
 		//get list of studentcourses
 		List<StudentCourse> studentcourses = st_cs_service.findAllCoursesByStudent(student.getStudentId());
-		//get ist of courses
-		List<Course> courseList = new ArrayList<Course>();
- 		for(StudentCourse stcourse : studentcourses) {
-			courseList.add(stcourse.getCourse());
-		}
-		//remove student from course
- 		for(Course course : courseList) {
- 			st_cs_service.removeStudentFromCourse(course, student);
+		//get list of courses
+//		List<Course> courseList = new ArrayList<Course>();
+// 		for(StudentCourse stcourse : studentcourses) {
+//			courseList.add(stcourse.getCourse());
+//		}
+ 		
+ 		for(StudentCourse sc: studentcourses)
+ 		{
+ 			st_cs_service.removeStudentFromCourse(sc.getCourse(), sc.getStudent());
  		}
  		
+		//remove student from course
+// 		for(Course course : courseList) {
+// 			st_cs_service.removeStudentFromCourse(course, student);
+// 		}
+ 		
  		//remove student
-		stservice.deleteStudent(stservice.findStudentByMatricNo(matricNo));
-		return "forward:/admin/studentlist";
+		stservice.deleteStudent(student);
+		return listStudent(model, session, true);
 		}
 	}
 	
 	
 	
 	@RequestMapping(value = "/lecturerlist")
-	public String listLecturer(Model model, HttpSession session) {
+	public String listLecturer(Model model, HttpSession session, Boolean delete_lecturer_status) {
 		Admin ad = (Admin) session.getAttribute("usession");
 		if (ad == null)
 		{
 			return "redirect:/home";
 		}
-		else {
-		model.addAttribute("lecturers", lservice.GetAllLecturers());
-		return "admin_lecturer_manage";
+		else
+		{
+			List<Lecturer> lecturers = lservice.GetAllLecturers();
+			
+			// status will be null when accessing this page for the first time (not coming from admin/deletelecturer)
+			if(delete_lecturer_status != null)
+			{
+				// If user is coming from admin/deletelecturer, show feedback whether delete is successful or not
+				model.addAttribute("status", delete_lecturer_status);
+			}
+			
+			model.addAttribute("lecturers", lecturers);
+		
+			return "admin_lecturer_manage";
 		}
 	}
 	
@@ -218,26 +240,33 @@ public class AdminController {
 	}
 	
 	@RequestMapping(value = "/deletelecturer/{lecturerId}")
-	public String deleteLecturer(@PathVariable("lecturerId") Integer lecturerId,  HttpSession session) {
+	public String deleteLecturer(@PathVariable("lecturerId") Integer lecturerId,  HttpSession session, Model model) {
 		Admin ad = (Admin) session.getAttribute("usession");
 		if (ad == null)
 		{
 			return "redirect:/home";
 		}
-		else {
-		List <Course> course_delete = cservice.findCoursesByLecturerId(lecturerId);
-		for(Course course: course_delete) {
-			course.setLecturer(null);
-			cservice.updateCourse(course);
-		}
-		lservice.deleteLecturer(lservice.findLecturerById(lecturerId));
-		return "forward:/admin/lecturerlist";
+		else 
+		{
+			List <Course> course_delete = cservice.findCoursesByLecturerId(lecturerId);
+			// Can only delete a lecturer when he has no more courses assigned to him
+			if (course_delete.size() == 0)
+			{
+				lservice.deleteLecturer(lservice.findLecturerById(lecturerId));
+				// true means delete is successful
+				return listLecturer(model, session, true);
+			}
+			else // There are still courses assigned to this lecturer, cannot delete lecturer
+			{
+				// false means delete is unsuccessful
+				return listLecturer(model, session, false);
+			}
 		}
 	}
 	
 	
 	@RequestMapping(value = "/enrolmentlist")
-	public String listEnrolment(Model model, Integer id, Integer statusId, HttpSession session)
+	public String listEnrolment(Model model, Integer id, Integer statusId, HttpSession session, Boolean enrol_student_course_status)
 	{
 		Admin ad = (Admin) session.getAttribute("usession");
 		if (ad == null)
@@ -288,18 +317,9 @@ public class AdminController {
 			model.addAttribute("selectedCourse", selectedCourse);
 		}
 		
-//		if(!status)
-//		{
-//			model.addAttribute("status", status);
-//		}
-//		Boolean status = (Boolean) model.getAttribute("status");
-		
-		if(statusId != null)
+		if(enrol_student_course_status != null)
 		{
-			if(statusId == 1)
-				model.addAttribute("status", true);
-			else
-				model.addAttribute("status", false);
+			model.addAttribute("status", enrol_student_course_status);
 		}
 		
 		return "admin_enrolment_manage";
@@ -322,14 +342,16 @@ public class AdminController {
 	@RequestMapping(value = "/enrolmentlist/{id}/{statusId}")
 	public String listEnrolment(@PathVariable("id") Integer id, Model model, @PathVariable("statusId") Integer statusId, HttpSession session)
 	{	
-		Admin ad = (Admin) session.getAttribute("usession");
-	if (ad == null)
-	{
-		return "redirect:/home";
-	}
-	else {
-		return listEnrolment(model, id, statusId, session);
-	}
+    Admin ad = (Admin) session.getAttribute("usession");
+    if (ad == null)
+	  {
+		  return "redirect:/home";
+    }
+    
+		if(statusId == 1)
+			return listEnrolment(model, id, session, true);
+		else
+			return listEnrolment(model, id, session, false);
 	}
 	
 //	@GetMapping("/addenrollment/{coursename}")
@@ -393,11 +415,11 @@ public class AdminController {
 		Course course = cservice.findCourseByCourseName(coursename);
 		//do assignment
 		boolean status = st_cs_service.adminAddStudentToCourse(course, student);
-		if(status)
+		if(status) // status == true means student has been added to course
 		{
 			return "redirect:/admin/enrolmentlist/{id}/1";
 		}
-		else
+		else // status == false means student could not be added to course
 		{
 			return "redirect:/admin/enrolmentlist/{id}/0";
 		}
@@ -424,15 +446,20 @@ public class AdminController {
 	
 	
 	@RequestMapping(value = "/courselist")
-	public String listCourse(Model model, HttpSession session) {
+	public String listCourse(Model model, HttpSession session, Boolean delete_course_status) {
 		Admin ad = (Admin) session.getAttribute("usession");
 		if (ad == null)
 		{
 			return "redirect:/home";
 		}
-		else {
-		model.addAttribute("courses", cservice.findAllCourseforCurrentYear());
-		return "admin_course_manage";
+		else 
+		{
+			List<Course> courses = cservice.findAllCourseforCurrentYear();
+			model.addAttribute("courses", courses);
+			if(delete_course_status != null)
+				model.addAttribute("status", delete_course_status);
+			
+			return "admin_course_manage";
 		}
 	}
 	
@@ -504,7 +531,7 @@ public class AdminController {
 	
 
 	@RequestMapping(value = "/deletecourse/{courseId}")
-	public String deleteCourse(@PathVariable("courseId") int courseId, HttpSession session) {
+	public String deleteCourse(@PathVariable("courseId") int courseId, HttpSession session, Model model) {
 	
 		Admin ad = (Admin) session.getAttribute("usession");
 		if (ad == null)
@@ -519,7 +546,7 @@ public class AdminController {
 		}
 		
 		cservice.deleteCourse(cservice.findCourseByCourseId(courseId));
-		return "forward:/admin/courselist";
+		return listCourse(model, session, true);
 		}
 	}
